Update README.md
Zuul is no more an awesome API gateway when we have Spring Cloud Gateway. Zuul is a nonreactive, outdated Gateway. Zuul-2 debatable can be on the list, but the Spring and Java community has decided to stay away from it and design the API gateway.. This later became the Spring Cloud Gateway.